Update Deploy-IntuneSystemtrayScript.ps1
Hi Jannik!

I do not intend this pull request to be a straight merge but I wanted to suggest some edits that I made for my usage. Modify or decline as you see fit, but I figured it was worth it to share!

The .ico I am using, I am pulling from my Azure blob storage since base64 encoding it comes out to about 55k characters. Because of this I opted for just a file transfer to the local system (C:\Windows\IntuneFiles rather than \Temp\). I used a .png of the current iteration of the Company Portal logo and it comes out really nice. (Example: https://i.imgur.com/OWc1irm.png)

Two other things I did was utilize $taskName for that random case that someone just wants to rename it, and during testing I put in some future proofing in case we modify the the script and need to redeploy, so I have it remove and re-add the task.

Your blog is great, thanks for you hard work!